Fix JPAN refactoring
