+++ src/main/java/org/scion/jpan/internal/MultiMap.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.scion.jpan;
+package org.scion.jpan.internal;
 
 import java.util.*;
