+++ src/main/java/org/scion/jpan/internal/Segments.java
@@ -22,7 +22,7 @@ import io.grpc.StatusRuntimeException;
 import java.nio.ByteBuffer;
 import java.time.Instant;
 import java.util.*;
-import org.scion.jpan.MultiMap;
+
 import org.scion.jpan.ScionRuntimeException;
 import org.scion.jpan.ScionUtil;
 import org.scion.jpan.proto.control_plane.Seg;
