+++ src/test/java/org/scion/jpan/internal/MultiMapTest.java
@@ -19,7 +19,6 @@ import static org.junit.jupiter.api.Assertions.*;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import org.scion.jpan.MultiMap;
 
 class MultiMapTest {
